Adds nvim-autopairs
nvim-autopairs is now used for pair completion in all files except tex.
For tex files we still use LuaSnip.

We change to auto-pairs because we don't want the LuaSnip tab behaviour
to exit outside of pairs. This is extremely annoying during usual
coding.

Plausibly a more intelligent LuaSnip config could fix this.